requirements.txt
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
APScheduler==3.10.4
asgiref==3.8.1
astor==0.8.1
attrs==23.2.0
bce-python-sdk==0.9.10
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
dashscope==1.19.1
dataclasses-json==0.6.6
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
Django==4.1.13
django-apscheduler==0.6.2
django-filter==23.2
django-ipware==6.0.5
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.14.0
filetype==1.2.0
fire==0.6.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface==0.0.1
huggingface-hub==0.23.1
idna==3.7
imageio==2.35.1
imgaug==0.4.0
inflection==0.5.1
intel-openmp==2021.4.0
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.7
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.4
langchain-huggingface==0.0.3
langchain-openai==0.0.8
langchain-text-splitters==0.0.2
langsmith==0.1.60
lazy_loader==0.4
lmdb==1.5.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
openai==1.30.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.3
packaging==23.2
paddleocr==2.8.1
paddlepaddle==2.6.2
pillow==10.3.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.19.6
psycopg2-binary==2.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-ipware==3.0.0
pytz==2024.1
PyYAML==6.0.1
qianfan==0.3.12
rapidfuzz==3.9.7
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.0
sentence-transformers==2.7.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tiktoken==0.5.2
tokenizers==0.19.1
torch==2.2.1
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.41.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
zhipuai==2.1.0.20240521
